requirements.txt
annotated-types==0.7.0
asgiref>=3.7,<4
dj-database-url==3.1.1
Django>=4.2,<6
django-ninja==1.5.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.11.0
gunicorn==21.2.0
django-cors-headers==4.3.1
openai>=1.0.0
httpx>=0.24.0
uvicorn[standard]
pytest>=7.4.0
pytest-django>=4.5.0
pytest-mock>=3.11.0
pytest-benchmark>=4.0.0
requests>=2.31.0
locust>=2.15.0
redis>=4.0.0